# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "host": "ec2-34-215-82-154.us-west-2.compute.amazonaws.com/laraveltest/public/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/auth/signup": {
            "post": {
                "tags": [
                    "Common"
                ],
                "description": "Registers a user, manufacturer and a technician",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Full Name of the user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email Id of the user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password of the user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user_type",
                        "in": "query",
                        "description": "The value of user_type could be one of ***user***, ***manufacturer*** and ***technician***",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "500": {
                        "description": "User registeration failed due to technical reasons"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "scopes": {
                "send:token": "Send the token"
            }
        }
    }
}